data_population_tool/requirements.txt
ipython==8.12.2
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
numpy==1.23.5
opencv_python_headless==4.7.0.72
pandas==2.0.2
Pillow==9.5.0
Pillow==10.0.0
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
